Added more things to evaluation little code refactorization and added some docstrings
